Move mainnet test into mainnet_config_test.go
