+++ config/params/config_test.go
@@ -3,7 +3,6 @@ package params_test
 import (
 	"bytes"
 	"math"
-	"path"
 	"sync"
 	"testing"
 
@@ -159,20 +158,3 @@ func TestMaxBlobsPerBlockByVersion(t *testing.T) {
 	}
 }
 
-func TestMainnetConfigMatchesUpstreamYaml(t *testing.T) {
-	presetFPs := presetsFilePath(t, "mainnet")
-	mn, err := params.ByName(params.MainnetName)
-	require.NoError(t, err)
-	cfg := mn.Copy()
-	for _, fp := range presetFPs {
-		cfg, err = params.UnmarshalConfigFile(fp, cfg)
-		require.NoError(t, err)
-	}
-	fPath, err := bazel.Runfile("external/mainnet")
-	require.NoError(t, err)
-	configFP := path.Join(fPath, "metadata", "config.yaml")
-	pcfg, err := params.UnmarshalConfigFile(configFP, nil)
-	require.NoError(t, err)
-	fields := fieldsFromYamls(t, append(presetFPs, configFP))
-	assertYamlFieldsMatch(t, "mainnet", fields, pcfg, params.BeaconConfig())
-}
