+++ config/params/mainnet_config_export_test.go
@@ -1,0 +1,6 @@
+package params
+
+// Re-exports for blackbox testing.
+const MainnetDenebForkEpoch = mainnetDenebForkEpoch
+var MainnetBeaconConfig = mainnetBeaconConfig
+
