+++ config/params/mainnet_config_test.go
@@ -1,9 +1,13 @@
-package params
+package params_test
 
 import (
+	"path"
 	"testing"
 
+	"github.com/OffchainLabs/prysm/v6/config/params"
 	"github.com/OffchainLabs/prysm/v6/consensus-types/primitives"
+	"github.com/OffchainLabs/prysm/v6/testing/require"
+	"github.com/bazelbuild/rules_go/go/tools/bazel"
 )
 
 func TestMaxRequestBlock(t *testing.T) {
@@ -13,21 +17,39 @@ func TestMaxRequestBlock(t *testing.T) {
 		description      string
 	}{
 		{
-			epoch:            primitives.Epoch(mainnetDenebForkEpoch - 1), // Assuming the fork epoch is not 0
-			expectedMaxBlock: mainnetBeaconConfig.MaxRequestBlocks,
+			epoch:            primitives.Epoch(params.MainnetDenebForkEpoch - 1), // Assuming the fork epoch is not 0
+			expectedMaxBlock: params.MainnetBeaconConfig.MaxRequestBlocks,
 		},
 		{
-			epoch:            primitives.Epoch(mainnetDenebForkEpoch),
-			expectedMaxBlock: mainnetBeaconConfig.MaxRequestBlocksDeneb,
+			epoch:            primitives.Epoch(params.MainnetDenebForkEpoch),
+			expectedMaxBlock: params.MainnetBeaconConfig.MaxRequestBlocksDeneb,
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.description, func(t *testing.T) {
-			maxBlocks := MaxRequestBlock(tc.epoch)
+			maxBlocks := params.MaxRequestBlock(tc.epoch)
 			if maxBlocks != tc.expectedMaxBlock {
 				t.Errorf("For epoch %d, expected max blocks %d, got %d", tc.epoch, tc.expectedMaxBlock, maxBlocks)
 			}
 		})
 	}
+}
+
+func TestMainnetConfigMatchesUpstreamYaml(t *testing.T) {
+	presetFPs := presetsFilePath(t, "mainnet")
+	mn, err := params.ByName(params.MainnetName)
+	require.NoError(t, err)
+	cfg := mn.Copy()
+	for _, fp := range presetFPs {
+		cfg, err = params.UnmarshalConfigFile(fp, cfg)
+		require.NoError(t, err)
+	}
+	fPath, err := bazel.Runfile("external/mainnet")
+	require.NoError(t, err)
+	configFP := path.Join(fPath, "metadata", "config.yaml")
+	pcfg, err := params.UnmarshalConfigFile(configFP, nil)
+	require.NoError(t, err)
+	fields := fieldsFromYamls(t, append(presetFPs, configFP))
+	assertYamlFieldsMatch(t, "mainnet", fields, pcfg, params.BeaconConfig())
 }
